Termiando as outras classes

+++ src/main/java/com/primeiroSistemaSpring/CadastroDeNinjas/model/Missoes.java
@@ -8,6 +8,16 @@ import lombok.Setter;
 
 import java.util.List;
 
+/**
+ * Entidade que representa uma Missão no sistema.
+ * <p>
+ * Esta classe mapeia a tabela tb_missoes no banco de dados e mantém
+ * informações sobre as missões disponíveis e os ninjas participantes.
+ * </p>
+ *
+ * @author Carlos
+ * @version 1.0
+ */
 @Entity
 @Table(name = "tb_missoes")
 @Getter
@@ -16,14 +26,30 @@ import java.util.List;
 @AllArgsConstructor
 public class Missoes {
 
+    /**
+     * Identificador único da missão.
+     * Gerado automaticamente pelo banco de dados.
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+    
+    /**
+     * Nome descritivo da missão.
+     */
     private String nomeMissao;
+    
+    /**
+     * Nível de dificuldade da missão (ex: Fácil, Média, Difícil).
+     */
     private String dificuldade;
 
-    // Uma missao pode ter varios ninjas
+    /**
+     * Lista de ninjas participantes desta missão.
+     * <p>
+     * Relacionamento One-to-Many: Uma missão pode ter vários ninjas.
+     * </p>
+     */
     @OneToMany(mappedBy = "missoes")
     private List<Ninja> ninjas;
-
 }
